add rate limiter for api calls, refactor requests creation
